Creating a separate file for the practical answers
The format of the ggplot part should now be similar to the rest of the course. The pdf has to be regenerated still. Not sure how you (@karlbrand / @elizabethribble ) do that, should I make a PDF myself and make a pull request, or will you regenerate the PDF from the Rmarkdown. (I would prefer the latter, as it makes sure the code and the PDF are always synchronized.